+++ client/src/navigation/screens/Friends.js
