+++ client/src/navigation/screens/Splash.js
