+++ client/src/navigation/screens/index.js
@@ -1,0 +1,11 @@
+export { default as Chat } from './ChatScreen/Chat';
+export { default as ChatMessage } from './ChatScreen/ChatMessage';
+export { default as Login } from './Login';
+export { default as Signup } from './Signup';
+export { default as Home } from './Home';
+export { default as Friends } from './Friends';
+export { default as Setting } from './Setting';
+export { default as NewsFeed } from './NewsFeed';
+export { default as Splash } from './Splash';
+export { default as DetailsUser } from './DetailsUser';
+export { default as Help } from './Help';
